Add get story handler

+++ src/API/routes/collection/service.ts
@@ -3,7 +3,6 @@ import { UserFacingError } from '../../errors/error'
 import { logger } from '../../../logging/logger'
 import { db } from '../../../db/database'
 import { hackernewsIntegration } from '../../../integrations/hackernews/hackernewsIntegration'
-import * as hackernewsIntegrationModule from '../../../integrations/hackernews/hackernewsIntegration'
 
 const addCommentTree = async (rootStoryId: string, commentId: number) => {
   const comment = await hackernewsIntegration.getComment(commentId)
@@ -13,34 +12,40 @@ const addCommentTree = async (rootStoryId: string, commentId: number) => {
     comment.kids.forEach((kid) => addCommentTree(rootStoryId, kid))
 }
 
-const handleCollectionAccess = async (userId: string, id: string) => {
+const validateCollectionAccess = async (userId: string, id: string) => {
   const collection = await db.collection.get(id)
-  if (collection.length === 0)
-    throw new UserFacingError(ErrorCode.collectionNotFound, 404)
-  if (collection[0].owner_id !== userId)
+  if (!collection) throw new UserFacingError(ErrorCode.collectionNotFound, 404)
+  if (collection.owner_id !== userId)
     throw new UserFacingError(ErrorCode.unauthorized, 401)
-  return collection
+  return
 }
 
 export const collectionService = {
-  get: (userId: string, id: string) => handleCollectionAccess(userId, id),
+  get: async (userId: string, id: string) => {
+    await validateCollectionAccess(userId, id)
+    return await db.collection.getWithStories(id)
+  },
   update: async (userId: string, id: string, name: string) => {
-    await handleCollectionAccess(userId, id)
+    await validateCollectionAccess(userId, id)
 
     return await db.collection.update(id, name)
   },
   new: (userId: string, name: string) => db.collection.save(userId, name),
   delete: async (userId: string, id: string) => {
-    await handleCollectionAccess(userId, id)
+    await validateCollectionAccess(userId, id)
 
     return await db.collection.delete(id)
   },
   add: async (userId: string, collectionId: string, storyId: number) => {
-    await handleCollectionAccess(userId, collectionId)
+    await validateCollectionAccess(userId, collectionId)
     const story = await hackernewsIntegration.getStory(storyId)
     const savedStory = (await db.item.saveStory(collectionId, story))[0]
     logger.debug(`Saved story ${story.id}`)
     story.kids &&
       story.kids.forEach((kid) => addCommentTree(savedStory.id, kid))
+  },
+  getStory: async (userId: string, collectionId: string, storyId: number) => {
+    await validateCollectionAccess(userId, collectionId)
+    return await db.item.getStoryWithComments(collectionId, storyId)
   }
 }
